Chore: Move product queries to inside the product component
