+++ react/components/ProductTranslation/graphql/getProduct.gql
