+++ react/components/ProductTranslation/graphql/startProductTranslations.gql
