+++ react/components/ProductTranslation/graphql/translateProduct.gql
