+++ react/components/ProductTranslation/graphql/uploadProductTranslation.gql
